restricting to own projects
this fixes the issue of assets being listed that are not owned by the current user, which could not be deleted, shows the number of submissions, and adds a filter for survey type assets.

+++ purge-all-assets.py
@@ -6,13 +6,15 @@ import requests
 
 URL = 'http://kf.kobotoolbox.org/api/v2/assets/?format=json'
 
-username = input('Username for "{}": '.format(URL))
+username = input(f'Username for "{URL}": ')
 password = getpass.getpass()
 auth = requests.auth.HTTPBasicAuth(username, password)
 
 response = requests.get(URL, auth=auth)
 for asset in response.json()['results']:
-    yn = input(f'Remove "{asset["name"]}" ({asset["uid"]})? (y/n) ')
-    if yn == 'y':
-        del_resp = requests.delete(asset['url'], auth=auth)
-        print(f'\t{del_resp.status_code}')
+    # limit to surveys owned by the current user
+    if asset['asset_type'] == 'survey' and asset['owner__username'] == username:
+        yn = input(f'Remove "{asset["name"]}" ({asset["deployment__submission_count"]} submissions)? (y/n) ')
+        if yn == 'y':
+            del_resp = requests.delete(asset['url'], auth=auth)
+            print(f'\t{del_resp.status_code}')
